+++ app/src/main/java/woowacourse/shopping/data/util/DoAsyncCatching.kt
@@ -1,0 +1,13 @@
+package woowacourse.shopping.data.util
+
+import kotlin.concurrent.thread
+
+inline fun <T> doAsyncCatching(
+    crossinline block: () -> T,
+    crossinline onResult: (Result<T>) -> Unit,
+) {
+    thread {
+        val result = runCatching { block() }
+        onResult(result)
+    }
+}
